factor out base NoaaHrrrRegionJob
